pyproject.toml
[tool.poetry]
name = "pd3r"
version = "0.1.0"
description = "Chat agent that writes federal position descriptions"
authors = ["you <you@example.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^1.0.6"
langgraph-checkpoint-sqlite = "^2.0"
langgraph-checkpoint-postgres = "^2.0"
pydantic = "^2.0"
pydantic-settings = "^2.0"
httpx = "^0.27"
jinja2 = "^3.1"
pillow = "^10.0"
langchain-openai = "^1.1.7"
# RAG dependencies
langchain-community = "^0.3"
langchain-text-splitters = "^0.3"
langchain-chroma = "^0.2"
pypdf = "^5.0"
chromadb = "^0.5"
# Export dependencies
python-docx = "^1.1"
langchain-anthropic = "^1.3.1"
# API dependencies
fastapi = "^0.115"
uvicorn = {version = "^0.34", extras = ["standard"]}
python-dotenv = "^1.0"
websockets = "^14.0"
python-multipart = "^0.0.18"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-m 'not llm'"
markers = [
    "llm: tests that require real LLM calls (deselected by default, run with: pytest -m llm)",
]

[tool.poetry.scripts]
pd3r = "src.main:main"
pd3r-api = "src.api.serve:main"

[tool.pd3r]
# Maximum number of draft elements to generate (0 = no limit)
max_drafts = 0

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
